:truck: Move `SentryBreadcrumbJsonProcessor` out of library config
